+++ messages/nl/base.php
@@ -8,7 +8,7 @@ return [
   'Hide users from these groups' => 'Gebruikers uit deze groepen verbergen',
   'Likes given' => 'Aantal "Vind ik leuk"',
   'Most Active Users Module Configuration' => 'Meest Actieve Gebruikers Module Configuratie',
-  'Posts created' => 'Berichten aangemaakt',
+  'Posts' => 'Berichten aangemaakt',
   'Save' => 'Bewaar',
   'The number of most active users that will be shown.' => 'Het aantal van de meest actieve gebruikers die getoond zullen worden.',
   'Total activity' => 'Totale activiteit',