+++ messages/pt-BR/base.php
@@ -8,7 +8,7 @@ return [
   'Hide users from these groups' => 'Ocultar usuários destes grupos',
   'Likes given' => 'Curtidas dadas',
   'Most Active Users Module Configuration' => 'Modulo de configuração de usuários mais ativos',
-  'Posts created' => 'Posts criados',
+  'Posts' => 'Posts criados',
   'Save' => 'Gravar.',
   'The number of most active users that will be shown.' => 'Número máximo de usuários mais ativos que será apresentado.',
   'Total activity' => 'Atividade total',