+++ widgets/Sidebar.php
@@ -9,7 +9,7 @@ class Sidebar extends \humhub\components\Widget
 {
     public function run()
     {
-        $users = ActiveUser::find()->limit((int) (new ConfigureForm())->noUsers)->all();
+        $users = ActiveUser::getMostActiveUsers((int) (new ConfigureForm())->noUsers);
         if (count($users) == 0) {
             return '';
         }