+++ widgets/views/sidebar.php
@@ -8,14 +8,9 @@ use humhub\widgets\PanelMenu;
 /**
  * @var $users ActiveUser[]
  */
-
-humhub\modules\mostactiveusers\assets\Assets::register($this);
 ?>
 <div class="panel panel-default" id="mostactiveusers-panel">
-
-    <!-- Display panel menu widget -->
-    <?php PanelMenu::widget(['id' => 'mostactiveusers-panel']); ?>
-
+    <?= PanelMenu::widget(['id' => 'mostactiveusers-panel']) ?>
     <div class="panel-heading">
         <?= Yii::t('MostactiveusersModule.base', '<strong>Most</strong> active people') ?>
     </div>